[ALPS03872830] wlan: Printf arg type mismatch:4272492
[Detail]
Printf arg type mismatch.

[Solution]
Add correct type casting for %u.

MTK-Commit-Id: 1041b8cdcfebfacccb73c539b6965c26dde8563d

CR-Id: ALPS03872830
Feature: Others

+++ os/linux/gl_proc.c
@@ -480,7 +480,7 @@ static ssize_t procDriverCmdRead(struct file *filp, char __user *buf,
 
 	if (u4CopySize > count) {
 		pr_err("count is too small: u4CopySize=%u, count=%u\n",
-		       u4CopySize, count);
+		       u4CopySize, (uint32_t)count);
 		return -EFAULT;
 	}
